# openapi.yaml
openapi: 3.0.3
info:
  title: 医养结合云端后台管理系统 API
  version: 2.0.0
  description: |
    基于 train.md 与当前 Spring Boot 实现维护的 REST API 契约。
    除登录、公开注册和刷新令牌外，接口均需要 Bearer access token。
    JSON 成功响应的业务 code 固定为 0，错误响应同时使用正确的 HTTP 状态码。
servers:
  - url: http://localhost:8080/api/v1
    description: 本地开发环境
security:
  - bearerAuth: []
...
paths:
...
  /elderly-profiles:
    get:
      tags: [ElderlyProfile]
      summary: 老人档案列表
      operationId: listElderlyProfiles
      parameters:
        - { $ref: '#/components/parameters/Keyword' }
        - { name: gender, in: query, schema: { type: string, enum: [male, female, unknown] } }
        - { name: careLevel, in: query, schema: { type: string } }
        - { name: status, in: query, schema: { type: string, enum: [active, inactive] } }
        - { name: regionCode, in: query, schema: { type: string } }
        - { $ref: '#/components/parameters/PageNo' }
        - { $ref: '#/components/parameters/PageSize' }
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiElderlyPage' }
    post:
      tags: [ElderlyProfile]
      summary: 新增老人档案
      operationId: createElderlyProfile
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ElderlyProfileCreateRequest' }
      responses:
        '201': { $ref: '#/components/responses/ElderlyResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
  /elderly-profiles/{id}:
    parameters:
      - { $ref: '#/components/parameters/Id' }
    get:
      tags: [ElderlyProfile]
      summary: 老人档案详情
      operationId: getElderlyProfile
      responses:
        '200': { $ref: '#/components/responses/ElderlyResponse' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [ElderlyProfile]
      summary: 更新老人档案
      operationId: updateElderlyProfile
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ElderlyProfileUpdateRequest' }
      responses:
        '200': { $ref: '#/components/responses/ElderlyResponse' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
    delete:
      tags: [ElderlyProfile]
      summary: 软删除老人档案
      operationId: deleteElderlyProfile
      responses:
        '200': { $ref: '#/components/responses/EmptyResponse' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /elderly-profiles/{id}/warnings:
    get:
      tags: [ElderlyProfile]
      summary: 查询老人关联预警
      operationId: getElderlyWarnings
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200': { $ref: '#/components/responses/WarningListResponse' }
        '404': { $ref: '#/components/responses/NotFound' }
  /elderly-profiles/{id}/reports:
    get:
      tags: [ElderlyProfile]
      summary: 查询老人关联报告
      operationId: getElderlyReports
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200': { $ref: '#/components/responses/AssessmentReportListResponse' }
        '404': { $ref: '#/components/responses/NotFound' }
  /elderly-profiles/{id}/devices:
    get:
      tags: [ElderlyProfile]
      summary: 查询老人关联设备
      operationId: getElderlyDevices
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200': { $ref: '#/components/responses/DeviceListResponse' }
        '404': { $ref: '#/components/responses/NotFound' }
  /elderly-profiles/{id}/key-populations:
    get:
      tags: [ElderlyProfile]
      summary: 查询老人重点人群记录
      operationId: getElderlyKeyPopulations
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200': { $ref: '#/components/responses/KeyPopulationListResponse' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /elderly-accounts:
    get:
      tags: [ElderlyAccount]
      summary: 老人账户列表
      operationId: listElderlyAccounts
      parameters:
        - { name: status, in: query, schema: { type: string, enum: [enabled, disabled] } }
        - { $ref: '#/components/parameters/PageNo' }
        - { $ref: '#/components/parameters/PageSize' }
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiUserPage' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [ElderlyAccount]
      summary: 创建老人账户
      operationId: createElderlyAccount
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SpecializedUserCreateRequest' }
      responses:
        '201': { $ref: '#/components/responses/UserResponse' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
  /elderly-accounts/{id}/reset-password:
    post:
      tags: [ElderlyAccount]
      summary: 重置老人账户密码
      operationId: resetElderlyPassword
      parameters: [{ $ref: '#/components/parameters/Id' }]
      requestBody:
        $ref: '#/components/requestBodies/Password'
      responses:
        '200': { $ref: '#/components/responses/EmptyResponse' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /elderly-accounts/{id}/status:
    post:
      tags: [ElderlyAccount]
      summary: 修改老人账户状态
      operationId: updateElderlyAccountStatus
      parameters: [{ $ref: '#/components/parameters/Id' }]
      requestBody:
        $ref: '#/components/requestBodies/UserStatus'
      responses:
        '200': { $ref: '#/components/responses/EmptyResponse' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /doctor-accounts:
    get:
      tags: [DoctorAccount]
      summary: 医生账户列表
      operationId: listDoctorAccounts
      parameters:
        - { name: status, in: query, schema: { type: string, enum: [enabled, disabled] } }
        - { $ref: '#/components/parameters/PageNo' }
        - { $ref: '#/components/parameters/PageSize' }
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiUserPage' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [DoctorAccount]
      summary: 创建医生账户
      operationId: createDoctorAccount
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SpecializedUserCreateRequest' }
      responses:
        '201': { $ref: '#/components/responses/UserResponse' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
  /doctor-accounts/{id}:
    put:
      tags: [DoctorAccount]
      summary: 更新医生账户
      operationId: updateDoctorAccount
      parameters: [{ $ref: '#/components/parameters/Id' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UserUpdateRequest' }
      responses:
        '200': { $ref: '#/components/responses/UserResponse' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /doctor-accounts/{id}/reset-password:
    post:
      tags: [DoctorAccount]
      summary: 重置医生账户密码
      operationId: resetDoctorPassword
      parameters: [{ $ref: '#/components/parameters/Id' }]
      requestBody:
        $ref: '#/components/requestBodies/Password'
      responses:
        '200': { $ref: '#/components/responses/EmptyResponse' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /users:
    get:
      tags: [UserAccount]
      summary: 系统账户列表
      operationId: listUsers
      parameters:
        - { $ref: '#/components/parameters/Keyword' }
        - { name: roleCode, in: query, schema: { $ref: '#/components/schemas/RoleCode' } }
        - { name: status, in: query, schema: { type: string, enum: [enabled, disabled] } }
        - { $ref: '#/components/parameters/PageNo' }
        - { $ref: '#/components/parameters/PageSize' }
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiUserPage' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [UserAccount]
      summary: 创建系统账户
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UserCreateRequest' }
      responses:
        '201': { $ref: '#/components/responses/UserResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
  /users/{id}:
    parameters:
      - { $ref: '#/components/parameters/Id' }
    get:
      tags: [UserAccount]
      summary: 系统账户详情
      operationId: getUser
      responses:
        '200': { $ref: '#/components/responses/UserResponse' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [UserAccount]
      summary: 更新系统账户
      operationId: updateUser
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UserUpdateRequest' }
      responses:
        '200': { $ref: '#/components/responses/UserResponse' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [UserAccount]
      summary: 软删除系统账户
      operationId: deleteUser
      responses:
        '200': { $ref: '#/components/responses/EmptyResponse' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: string, maxLength: 64 }
    Keyword:
      name: keyword
      in: query
      schema: { type: string, maxLength: 100 }
    PageNo:
      name: pageNo
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      name: pageSize
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 10 }
  requestBodies:
...
    Password:
      required: true
      content:
        application/json:
          schema: { $ref: '#/components/schemas/PasswordRequest' }
    UserStatus:
      required: true
      content:
        application/json:
          schema: { $ref: '#/components/schemas/UserStatusUpdateRequest' }
  responses:
    EmptyResponse:
      description: 成功
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiEmpty' }
...
    UserResponse:
      description: 成功
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiUser' }
    ElderlyResponse:
      description: 成功
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiElderly' }
...
    WarningListResponse:
      description: 成功
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiWarningList' }
...
    AssessmentReportListResponse:
      description: 成功
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiAssessmentReportList' }
...
    KeyPopulationListResponse:
      description: 成功
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiKeyPopulationList' }
...
    DeviceListResponse:
      description: 成功
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiDeviceList' }
...
    BadRequest:
      description: 参数或业务输入错误
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
...
    Forbidden:
      description: 当前角色无权访问
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    NotFound:
      description: 数据不存在
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    Conflict:
      description: 数据冲突或仍被引用
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
  schemas:
    ApiBase:
      type: object
      required: [code, message, traceId]
      properties:
        code: { type: integer, example: 0 }
        message: { type: string, example: success }
        traceId: { type: string, example: b3c1f8e9d8f14a7f }
    ApiEmpty:
      allOf:
        - { $ref: '#/components/schemas/ApiBase' }
        - type: object
          properties:
            data: { nullable: true }
    ApiError:
      type: object
      required: [code, message, traceId]
      properties:
        code: { type: integer, enum: [400, 401, 403, 404, 409, 500, 501] }
        message: { type: string }
        data: { nullable: true }
        traceId: { type: string }
...
    RoleCode:
      type: string
      enum: [admin, doctor, operator, analyst, elderly]
    User:
      type: object
      properties:
        id: { type: string }
        username: { type: string }
        realName: { type: string }
        roleCode: { $ref: '#/components/schemas/RoleCode' }
        mobile: { type: string, nullable: true }
        institutionId: { type: string, nullable: true }
        regionCode: { type: string, nullable: true }
        status: { type: string, enum: [enabled, disabled] }
        lastLoginAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    UserCreateRequest:
      type: object
      required: [username, password, realName, roleCode]
      properties:
        username: { type: string, minLength: 3, maxLength: 50, pattern: '^[A-Za-z0-9_]+$' }
        password: { type: string, minLength: 8, maxLength: 64, writeOnly: true }
        realName: { type: string, maxLength: 50 }
        roleCode: { $ref: '#/components/schemas/RoleCode' }
        mobile: { type: string, pattern: '^1[3-9]\d{9}$' }
        institutionId: { type: string, maxLength: 64 }
        regionCode: { type: string, maxLength: 64 }
        status: { type: string, enum: [enabled, disabled], default: enabled }
    SpecializedUserCreateRequest:
      type: object
      required: [username, password, realName]
      properties:
        username: { type: string, minLength: 3, maxLength: 50, pattern: '^[A-Za-z0-9_]+$' }
        password: { type: string, minLength: 8, maxLength: 64, writeOnly: true }
        realName: { type: string, maxLength: 50 }
        mobile: { type: string, pattern: '^1[3-9]\d{9}$' }
        institutionId: { type: string, maxLength: 64 }
        regionCode: { type: string, maxLength: 64 }
        status: { type: string, enum: [enabled, disabled], default: enabled }
    UserUpdateRequest:
      type: object
      properties:
        realName: { type: string, maxLength: 50 }
        roleCode: { $ref: '#/components/schemas/RoleCode' }
        mobile: { type: string, pattern: '^1[3-9]\d{9}$' }
        institutionId: { type: string, maxLength: 64 }
        regionCode: { type: string, maxLength: 64 }
        status: { type: string, enum: [enabled, disabled] }
    UserStatusUpdateRequest:
      type: object
      required: [status]
      properties:
        status: { type: string, enum: [enabled, disabled] }
...
    PasswordRequest:
      type: object
      required: [newPassword]
      properties:
        newPassword: { type: string, minLength: 8, maxLength: 64, writeOnly: true }
...
    ApiUser:
      allOf:
        - { $ref: '#/components/schemas/ApiBase' }
        - type: object
          properties:
            data: { $ref: '#/components/schemas/User' }
    UserPage:
      type: object
      required: [list, pageNo, pageSize, total]
      properties:
        list: { type: array, items: { $ref: '#/components/schemas/User' } }
        pageNo: { type: integer }
        pageSize: { type: integer }
        total: { type: integer, format: int64 }
    ApiUserPage:
      allOf:
        - { $ref: '#/components/schemas/ApiBase' }
        - type: object
          properties:
            data: { $ref: '#/components/schemas/UserPage' }
    ElderlyProfile:
      type: object
      properties:
        id: { type: string }
        userId: { type: string, nullable: true }
        name: { type: string }
        gender: { type: string, enum: [male, female, unknown] }
        birthday: { type: string, format: date, nullable: true }
        age: { type: integer, nullable: true }
        phone: { type: string, nullable: true }
        address: { type: string, nullable: true }
        institutionId: { type: string, nullable: true }
        regionCode: { type: string, nullable: true }
        medicalHistory: { type: string, nullable: true }
        careLevel: { type: string, nullable: true }
        tags: { type: array, items: { type: string } }
        status: { type: string, enum: [active, inactive] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    ElderlyProfileCreateRequest:
      type: object
      required: [name, gender]
      properties:
        id: { type: string, maxLength: 64 }
        userId: { type: string, maxLength: 64 }
        name: { type: string, maxLength: 100 }
        gender: { type: string, enum: [male, female, unknown] }
        birthday: { type: string, format: date }
        age: { type: integer, minimum: 0 }
        phone: { type: string, maxLength: 20 }
        address: { type: string, maxLength: 500 }
        institutionId: { type: string, maxLength: 64 }
        regionCode: { type: string, maxLength: 64 }
        medicalHistory: { type: string, maxLength: 1000 }
        careLevel: { type: string, maxLength: 10 }
        tags: { type: array, items: { type: string, maxLength: 50 } }
        status: { type: string, enum: [active, inactive], default: active }
    ElderlyProfileUpdateRequest:
      type: object
      properties:
        userId: { type: string, maxLength: 64 }
        name: { type: string, maxLength: 100 }
        gender: { type: string, enum: [male, female, unknown] }
        birthday: { type: string, format: date }
        age: { type: integer, minimum: 0 }
        phone: { type: string, maxLength: 20 }
        address: { type: string, maxLength: 500 }
        institutionId: { type: string, maxLength: 64 }
        regionCode: { type: string, maxLength: 64 }
        medicalHistory: { type: string, maxLength: 1000 }
        careLevel: { type: string, maxLength: 10 }
        tags: { type: array, items: { type: string, maxLength: 50 } }
        status: { type: string, enum: [active, inactive] }
    ApiElderly:
      allOf:
        - { $ref: '#/components/schemas/ApiBase' }
        - type: object
          properties:
            data: { $ref: '#/components/schemas/ElderlyProfile' }
    ElderlyPage:
      type: object
      required: [list, pageNo, pageSize, total]
      properties:
        list: { type: array, items: { $ref: '#/components/schemas/ElderlyProfile' } }
        pageNo: { type: integer }
        pageSize: { type: integer }
        total: { type: integer, format: int64 }
    ApiElderlyPage:
      allOf:
        - { $ref: '#/components/schemas/ApiBase' }
        - type: object
          properties:
            data: { $ref: '#/components/schemas/ElderlyPage' }
    HealthWarning:
      type: object
      properties:
        id: { type: string }
        elderlyId: { type: string }
        warningType: { type: string }
        severity: { type: string, enum: [low, medium, high, critical] }
        source: { type: string }
        metricName: { type: string, nullable: true }
        metricValue: { type: number, format: double, nullable: true }
        thresholdValue: { type: number, format: double, nullable: true }
        status: { type: string, enum: [unprocessed, processing, processed, closed] }
        occurredAt: { type: string, format: date-time }
        handledAt: { type: string, format: date-time, nullable: true }
        handlerId: { type: string, nullable: true }
        handlerName: { type: string, nullable: true }
        handleResult: { type: string, nullable: true }
        nextAction: { type: string, nullable: true }
        remark: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    ApiWarningList:
      allOf:
        - { $ref: '#/components/schemas/ApiBase' }
        - type: object
          properties:
            data: { type: array, items: { $ref: '#/components/schemas/HealthWarning' } }
...
    AssessmentReport:
      type: object
      properties:
        id: { type: string }
        elderlyId: { type: string }
        reportType: { type: string }
        score: { type: integer, minimum: 0, maximum: 100 }
        grade: { type: string }
        summary: { type: string }
        riskItems: { type: array, items: { type: string } }
        recommendations: { type: array, items: { type: string } }
        assessorId: { type: string, nullable: true }
        assessedAt: { type: string, format: date-time }
        reviewStatus: { type: string, enum: [draft, approved, rejected] }
        reviewerId: { type: string, nullable: true }
        reviewedAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    ApiAssessmentReportList:
      allOf:
        - { $ref: '#/components/schemas/ApiBase' }
        - type: object
          properties:
            data: { type: array, items: { $ref: '#/components/schemas/AssessmentReport' } }
...
    KeyPopulation:
      type: object
      properties:
        id: { type: string }
        elderlyId: { type: string }
        category: { type: string }
        reason: { type: string, nullable: true }
        level: { type: string }
        ownerDoctorId: { type: string, nullable: true }
        followUpCycleDays: { type: integer }
        status: { type: string, enum: [active, closed] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    ApiKeyPopulationList:
      allOf:
        - { $ref: '#/components/schemas/ApiBase' }
        - type: object
          properties:
            data: { type: array, items: { $ref: '#/components/schemas/KeyPopulation' } }
...
    Device:
      type: object
      properties:
        id: { type: string }
        deviceName: { type: string }
        deviceType: { type: string }
        deviceSn: { type: string }
        elderlyId: { type: string, nullable: true }
        bindingStatus: { type: string, enum: [bound, unbound] }
        onlineStatus: { type: string, enum: [online, offline] }
        lastReportAt: { type: string, format: date-time, nullable: true }
        firmwareVersion: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    ApiDeviceList:
      allOf:
        - { $ref: '#/components/schemas/ApiBase' }
        - type: object
          properties:
            data: { type: array, items: { $ref: '#/components/schemas/Device' } }
...